Fix issue #13

+++ ml_insights/insights.py
@@ -223,7 +223,7 @@ class ModelXRay(object):
             -------
         '''
         ## Convert Pandas DataFrame to nparray explicitly to make life easier
-        
+
         columns = list(self.results.keys())
         result_data = [importance_distribution_of_variable(self.results[col_name][1]) for col_name in columns]
         sortind = np.argsort([np.median(d) for d in result_data])
@@ -237,7 +237,7 @@ class ModelXRay(object):
             ax = _gca()
             fig = ax.get_figure()
             fig.set_figwidth(10)
-            fig.set_figheight(max(6, math.ceil(num_features*0.5)))
+            fig.set_figheight(max(6, int(math.ceil(num_features*0.5))))
         ax.boxplot(plot_data, notch=0, sym='+', vert=0, whis=1.5)
         ax.set_yticklabels([columns[idx] for idx in sortind][-num_features:]);
 
@@ -246,11 +246,11 @@ class ModelXRay(object):
         '''This function visualizes the effect of a single variable in models with complicated dependencies.
         Given a dataset, it will select points in that dataset, and then change the select column across
         different values to view the effect of the model prediction given that variable. These have been called
-        Individual Conditional Expectation plots (or ICE-plots), see Goldstein, Kapelner, Bleich, 
-        Pitkin. Peeking Inside the Black Box: Visualizing Statistical Learning With Plots of Individual 
+        Individual Conditional Expectation plots (or ICE-plots), see Goldstein, Kapelner, Bleich,
+        Pitkin. Peeking Inside the Black Box: Visualizing Statistical Learning With Plots of Individual
         Conditional Expectation. Journal of Computational and Graphical Statistics (March 2014)
         '''
-        
+
         import matplotlib.pyplot as plt
 
         columns = sorted(list(self.results.keys()))
@@ -267,7 +267,7 @@ class ModelXRay(object):
             y_base_points = None
 
         n_cols = min(3, len(columns))
-        n_rows = math.ceil(len(columns) / n_cols)
+        n_rows = int(math.ceil(len(columns) / n_cols))
         figsize = (n_cols * 4, n_rows * 4)
         fig, axes = plt.subplots(n_rows, n_cols, figsize=figsize)
         for col_name, ax in zip(columns, axes.flatten()):
@@ -287,10 +287,10 @@ class ModelXRay(object):
         '''Given the indices of two points in the "xray"-ed data set, this function gives an explanation
         of the factors contributing to the difference in the predictions.
 
-        Starting with the first point given, the considers changing each feature from its current value to that 
+        Starting with the first point given, the considers changing each feature from its current value to that
         possessed by the second point.  The function evaluates the target in both scenarios and determines the
-        feature value change that creates the biggest (absolute) change in the target.  This change is selected 
-        and the current point becomes the new point with the new feature value.  This is repeated until the new 
+        feature value change that creates the biggest (absolute) change in the target.  This change is selected
+        and the current point becomes the new point with the new feature value.  This is repeated until the new
         target value is within a factor of 1+tol of the second point.
         '''
         data_row_1 = self._get_data_rows(index_1)
@@ -308,10 +308,10 @@ def explain_prediction_difference(model, data_row_1, data_row_2, tol=.03, verbos
     '''Given a model and two single row data frames, this function gives an explanation
         of the factors contributing to the difference in the predictions.
 
-        Starting with the first point given, the considers changing each feature from its current value to that 
+        Starting with the first point given, the considers changing each feature from its current value to that
         possessed by the second point.  The function evaluates the target in both scenarios and determines the
-        feature value change that creates the biggest (absolute) change in the target.  This change is selected 
-        and the current point becomes the new point with the new feature value.  This is repeated until the new 
+        feature value change that creates the biggest (absolute) change in the target.  This change is selected
+        and the current point becomes the new point with the new feature value.  This is repeated until the new
         target value is within a factor of 1+tol of the second point.
         '''
     column_names = data_row_1.index
